+++ src/main/java/com/yanny/ytech/configuration/block/AqueductConsumerBlock.java
@@ -34,7 +34,6 @@ public abstract class AqueductConsumerBlock extends IrrigationBlock {
         super(Properties.ofFullCopy(Blocks.TERRACOTTA));
     }
 
-    @SuppressWarnings("deprecation")
     @NotNull
     @Override
     public RenderShape getRenderShape(@NotNull BlockState blockState) {