+++ src/main/java/com/yanny/ytech/configuration/block/MillstoneBlock.java
@@ -3,6 +3,7 @@ package com.yanny.ytech.configuration.block;
 import com.yanny.ytech.configuration.Utils;
 import com.yanny.ytech.configuration.block_entity.MillstoneBlockEntity;
 import com.yanny.ytech.configuration.recipe.WorkspaceCraftingRecipe;
+import com.yanny.ytech.registration.YTechBlockEntityTypes;
 import com.yanny.ytech.registration.YTechBlocks;
 import com.yanny.ytech.registration.YTechItemTags;
 import net.minecraft.core.BlockPos;
@@ -34,6 +35,8 @@ import net.neoforged.neoforge.client.model.generators.ConfiguredModel;
 import net.neoforged.neoforge.client.model.generators.ModelFile;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
+
+import java.util.Optional;
 
 public class MillstoneBlock extends Block implements EntityBlock {
     private static final VoxelShape SHAPE = Shapes.or(
@@ -70,11 +73,13 @@ public class MillstoneBlock extends Block implements EntityBlock {
     @Override
     public ItemInteractionResult useItemOn(@NotNull ItemStack stack, @NotNull BlockState state, @NotNull Level level, @NotNull BlockPos pos, @NotNull Player player,
                                            @NotNull InteractionHand hand, @NotNull BlockHitResult hitResult) {
-        if (level.getBlockEntity(pos) instanceof  MillstoneBlockEntity millstone) {
-            return millstone.onUse(level, pos, player, hand);
-        } else {
-            throw new IllegalStateException("Invalid holder type!");
+        Optional<MillstoneBlockEntity> optional = level.getBlockEntity(pos, YTechBlockEntityTypes.MILLSTONE.get());
+
+        if (optional.isPresent()) {
+            return optional.get().onUse(level, pos, player, hand);
         }
+
+        return ItemInteractionResult.PASS_TO_DEFAULT_BLOCK_INTERACTION;
     }
 
     @Override