+++ src/main/java/com/yanny/ytech/configuration/block/WellPulleyBlock.java
@@ -2,6 +2,7 @@ package com.yanny.ytech.configuration.block;
 
 import com.yanny.ytech.configuration.Utils;
 import com.yanny.ytech.configuration.block_entity.WellPulleyBlockEntity;
+import com.yanny.ytech.registration.YTechBlockEntityTypes;
 import com.yanny.ytech.registration.YTechBlocks;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
@@ -42,6 +43,7 @@ import org.jetbrains.annotations.Nullable;
 
 import java.util.List;
 import java.util.Objects;
+import java.util.Optional;
 
 import static net.minecraft.world.level.block.state.properties.BlockStateProperties.*;
 
@@ -57,7 +59,6 @@ public class WellPulleyBlock extends IrrigationBlock {
         super(Properties.ofFullCopy(Blocks.TERRACOTTA).pushReaction(PushReaction.DESTROY));
     }
 
-    @SuppressWarnings("deprecation")
     @NotNull
     @Override
     public VoxelShape getShape(BlockState pState, @NotNull BlockGetter pLevel, @NotNull BlockPos pPos, @NotNull CollisionContext pContext) {
@@ -87,11 +88,13 @@ public class WellPulleyBlock extends IrrigationBlock {
             aboveState = level.getBlockState(basePos.above());
         }
 
-        if (level.getBlockEntity(basePos) instanceof WellPulleyBlockEntity blockEntity) {
-            return blockEntity.onUse(level, aboveState, player);
-        } else {
-            throw new IllegalStateException("Invalid holder type!");
+        Optional<WellPulleyBlockEntity> optional = level.getBlockEntity(pos, YTechBlockEntityTypes.WELL_PULLEY.get());
+
+        if (optional.isPresent()) {
+            return optional.get().onUse(level, aboveState, player);
         }
+
+        return ItemInteractionResult.PASS_TO_DEFAULT_BLOCK_INTERACTION;
     }
 
     @NotNull
@@ -153,7 +156,6 @@ public class WellPulleyBlock extends IrrigationBlock {
         return null;
     }
 
-    @SuppressWarnings("deprecation")
     @NotNull
     @Override
     public BlockState updateShape(@NotNull BlockState state, @NotNull Direction direction, @NotNull BlockState neighborState,