+++ src/main/java/com/yanny/ytech/configuration/block_entity/AqueductBlockEntity.java
@@ -10,7 +10,10 @@ import net.minecraft.core.HolderLookup;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.level.biome.Biome;
+import net.minecraft.world.level.block.PointedDripstoneBlock;
 import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.level.material.Fluid;
+import net.minecraft.world.level.material.Fluids;
 import org.jetbrains.annotations.NotNull;
 
 public class AqueductBlockEntity extends IrrigationBlockEntity {
@@ -42,6 +45,20 @@ public class AqueductBlockEntity extends IrrigationBlockEntity {
             Holder<Biome> biome = level.getBiome(worldPosition);
             return !YTechMod.CONFIGURATION.isValidBlockForRaining() || (level.canSeeSky(worldPosition.above())
                     && biome.value().getPrecipitationAt(worldPosition) == Biome.Precipitation.RAIN);
+        }
+
+        return false;
+    }
+
+    @Override
+    public boolean validForDripping() {
+        if (level instanceof ServerLevel serverLevel) {
+            BlockPos pos = PointedDripstoneBlock.findStalactiteTipAboveCauldron(serverLevel, worldPosition);
+
+            if (pos != null) {
+                Fluid fluid = PointedDripstoneBlock.getCauldronFillFluidType(serverLevel, pos);
+                return fluid != Fluids.EMPTY && fluid.isSame(Fluids.WATER);
+            }
         }
 
         return false;